changes from main

+++ cmput404-group-project/backend/social_distribution/service/views.py
@@ -1,9 +1,12 @@
+from itertools import chain
 from django.shortcuts import get_object_or_404, render
 from rest_framework import viewsets, status, generics, mixins
 from rest_framework.decorators import api_view, action
 from rest_framework.response import Response
 from django.contrib.auth import authenticate, login
 from django.db.models import Q
+from rest_framework.permissions import AllowAny
+from rest_framework.authtoken.models import Token
 
 from .paginations import PostsPagination, CommentsPagination
 
@@ -33,6 +36,7 @@ baseURL = "http://127.0.0.1:8000/"
 class UsersViewSet(viewsets.GenericViewSet):
     queryset = User.objects.all()
     serializer_class = UserSerializer
+    permission_classes = [AllowAny]
     
     def create(self, request, *args, **kwargs):
         # create corresponding author
@@ -59,30 +63,49 @@ class UsersViewSet(viewsets.GenericViewSet):
         liked = Liked(id=liked_id, author=author)
         liked.save()
 
+        token = Token.objects.create(user=user)
+
         author_serializer = AuthorSerializer(instance=author)
 
-        return Response(author_serializer.data, status=status.HTTP_201_CREATED)
+        return Response(
+            {
+                'author': author_serializer.data,
+                'token': token.key,
+            },
+            status=status.HTTP_201_CREATED)
     
-    def list(self, request):
-        recent_users = User.objects.all().order_by('-last_login')
-        # page = self.paginate_queryset(recent_users)
-        # if page is not None:
-        #     serializer = self.get_serializer(page, many=True)
-        #     return self.get_paginated_response(serializer.data)
-        serializer = self.get_serializer(recent_users, many=True)
-        return Response(serializer.data)
+    @action(detail=False)
+    def auth_test(self, request, *args, **kwargs):
+        content = {
+            'user': str(request.user),  # `django.contrib.auth.User` instance.
+            'auth': str(request.auth),  # None
+        }
+        return Response(content)
+    
+    ####### delete this
+    # def list(self, request):
+    #     recent_users = User.objects.all().order_by('-last_login')
+    #     # page = self.paginate_queryset(recent_users)
+    #     # if page is not None:
+    #     #     serializer = self.get_serializer(page, many=True)
+    #     #     return self.get_paginated_response(serializer.data)
+    #     serializer = self.get_serializer(recent_users, many=True)
+    #     return Response(serializer.data)
 
     @action(detail=True, methods=['post'])
     def update_pass(self, request, *args, **kwargs):
-        self.object = self.get_object()
+        object = self.get_object()
+        if request.user != object:
+            return Response({"detail:" ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
         serializer = ChangePasswordSerializer(data=request.data)
 
         if serializer.is_valid():
-            if not self.object.check_password(serializer.data.get("old_password")):
+            if not object.check_password(serializer.data.get("old_password")):
                 return Response({"old_password": ["Wrong Password."]}, 
                                 status=status.HTTP_401_UNAUTHORIZED)
-            self.object.set_password(serializer.data.get("new_password"))
-            self.object.save()
+            object.set_password(serializer.data.get("new_password"))
+            object.save()
             return Response({"status": "success",
                              "code": status.HTTP_200_OK,
                              "message": "Password updated successfully"})
@@ -100,7 +123,11 @@ class UsersViewSet(viewsets.GenericViewSet):
                 login(request, user)
                 author = Author.get_author_from_user(user=user)
                 author_serializer = AuthorSerializer(instance=author)
-                return Response(author_serializer.data,
+                token = get_object_or_404(Token, user=user)
+                return Response({
+                    'author': author_serializer.data,
+                    'token': token.key,
+                },
                         status=status.HTTP_200_OK)
             return Response({"detail": ["Wrong username or password."]}, 
                             status=status.HTTP_401_UNAUTHORIZED)
@@ -108,15 +135,50 @@ class UsersViewSet(viewsets.GenericViewSet):
                         status=status.HTTP_400_BAD_REQUEST)
 
 
-class AuthorsViewSet(viewsets.ModelViewSet):
+class AuthorsViewSet(viewsets.GenericViewSet):
     queryset = Author.objects.all()
     serializer_class = AuthorSerializer
+    # permission_classes = [IsAuthenticated]
 
     def list(self, request, *args, **kwargs): # overrides the default list method
         authors = Author.objects.all()
         serializer = self.get_serializer(authors, many=True)
         return Response({"type": "authors",
                          "items": serializer.data})
+    
+    def retrieve(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.get_serializer(instance)
+        return Response(serializer.data)
+    
+    def update(self, request, *args, **kwargs):
+        partial = kwargs.pop('partial', False)
+        user = request.user
+        instance = self.get_object()
+        if (instance.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        serializer = self.get_serializer(instance, data=request.data, partial=partial)
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+
+        if getattr(instance, '_prefetched_objects_cache', None):
+            # If 'prefetch_related' has been applied to a queryset, we need to
+            # forcibly invalidate the prefetch cache on the instance.
+            instance._prefetched_objects_cache = {}
+
+        return Response(serializer.data)
+    
+    @action(detail=True)
+    def get_user(self, request, pk, *args, **kwargs):
+        user = request.user
+        instance = self.get_object()
+        if (instance.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        user = instance.user
+        serializer = UserSerializer(instance=user)
+        return Response(serializer.data)
     
     @action(detail=True)
     def friends(self, request, pk, *args, **kwargs):
@@ -174,23 +236,35 @@ class FollowRequestViewSet(viewsets.GenericViewSet):
             return Response({"detail": ["Request does not exist."]},
                             status=status.HTTP_404_NOT_FOUND)
 
-    @action(detail=True, methods=['get', 'post'])
+    @action(detail=True, methods=['post'])
     def send(self, request, pk=None, author_pk=None, *args, **kwargs):
+        user = request.user
         object = Author.objects.get(id=author_pk)
         actor = Author.objects.get(id=pk)
+        if (actor.user != user): # user must be the actor
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        
         summary = f"{actor.displayName} wants to follow {object.displayName}."
         if FollowRequest.objects.filter(actor=actor, object=object).count(): # request already exists
             return Response({"detail": ["Request already exists."]},
                             status=status.HTTP_400_BAD_REQUEST)
-        follow_request = FollowRequest(summary=summary, actor=actor, object=object)
+        id = baseURL + "service/authors/"+object.id+"/follow-request"+actor.id+"/"
+        follow_request = FollowRequest(id=id, summary=summary, actor=actor, object=object)
         follow_request.save()
         serializer = FollowRequestSerializer(instance=follow_request)
         return Response(serializer.data)
     
-    @action(detail=True, methods=['get', 'post'])
+    @action(detail=True, methods=['post'])
     def accept(self, request, pk=None, author_pk=None, *args, **kwargs):
+        user = request.user
         object = get_object_or_404(Author, id=author_pk)
         actor = get_object_or_404(Author, id=pk)
+
+        if (object.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+
         if FollowRequest.objects.filter(actor=actor, object=object).count():
             instance = FollowRequest.objects.get(object=object, actor=actor)
             followers = Followers.objects.get(author=object)
@@ -204,10 +278,16 @@ class FollowRequestViewSet(viewsets.GenericViewSet):
         return Response({"detail": ["Request does not exist."]},
                         status=status.HTTP_404_NOT_FOUND)
     
-    @action(detail=True, methods=['get', 'post'])
+    @action(detail=True, methods=['post'])
     def decline(self, request, pk=None, author_pk=None, *args, **kwargs):
+        user = request.user
         object = get_object_or_404(Author, id=author_pk)
         actor = get_object_or_404(Author, id=pk)
+
+        if (object.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+
         if FollowRequest.objects.filter(actor=actor, object=object).count():
             instance = FollowRequest.objects.get(object=object, actor=actor)
             instance.delete()
@@ -230,25 +310,20 @@ class FollowersViewSet(viewsets.GenericViewSet):
     @action(detail=False, methods=['post'])
     def unfollow(self, request, author_pk=None, *args, **kwargs):
         author = get_object_or_404(Author, id=author_pk)
-        serializer = AuthorIdSerializer(data=request.data)
-        if serializer.is_valid():
-            pk = serializer.data.get("id")
-            follower = get_object_or_404(Author, id=pk)
-            followers = Followers.objects.get(author=author)
-            following = Following.objects.get(author=follower)
-            if followers.items.contains(follower) & following.items.contains(author):
-                followers.items.remove(follower)
-                following.items.remove(author)
-                return Response({"detail": ["Unfollowed successfully."]},
-                                status=status.HTTP_200_OK)
+        user = request.user
+        follower = get_object_or_404(Author, user=user)
+        followers = Followers.objects.get(author=author)
+        following = Following.objects.get(author=follower)
+        if followers.items.contains(follower) & following.items.contains(author):
+            followers.items.remove(follower)
+            following.items.remove(author)
+            return Response({"detail": ["Unfollowed successfully."]},
+                            status=status.HTTP_200_OK)
 
-            return Response({"detail": ["User is not in following list."]},
-                            status=status.HTTP_404_NOT_FOUND)
-        
-        return Response(serializer.errors,
-                        status=status.HTTP_400_BAD_REQUEST)
+        return Response({"detail": ["User is not in following list."]},
+                        status=status.HTTP_404_NOT_FOUND)
 
-class FollowingViewSet(viewsets.ModelViewSet):
+class FollowingViewSet(viewsets.GenericViewSet):
     queryset = Following.objects.all()
     serializer_class = FollowingSerializer
 
@@ -258,19 +333,24 @@ class FollowingViewSet(viewsets.ModelViewSet):
         return Response(serializer.data)
 
 
-class PostsViewSet(viewsets.ModelViewSet):
+class PostsViewSet(viewsets.GenericViewSet):
     queryset = Post.objects.all()
     serializer_class = PostSerializer
-    # pagination_class = PostsPagination
+    pagination_class = PostsPagination
 
     def list(self, request, author_pk=None, *args, **kwargs): # overrides the default list method
         posts = Post.objects.filter(author__id = author_pk).all()
-        serializer = self.get_serializer(posts, many=True)
-        return Response({"type": "posts",
-                         "items": serializer.data})
+        page = self.paginate_queryset(posts)
+        serializer = self.get_serializer(page, many=True)
+        return self.get_paginated_response(serializer.data)
     
     def create(self, request, author_pk=None, *args, **kwargs):
+        user = request.user
         author = Author.objects.get(id=author_pk)
+        if (author.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         post = serializer.save(author=author)
@@ -282,7 +362,12 @@ class PostsViewSet(viewsets.ModelViewSet):
         headers = self.get_success_headers(serializer.data)
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
     
-    def update(self, request, *args, **kwargs):
+    def update(self, request, author_pk=None, *args, **kwargs):
+        user = request.user
+        author = Author.objects.get(id=author_pk)
+        if (author.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
         partial = kwargs.pop('partial', False)
         instance = self.get_object()
         serializer = self.get_serializer(instance, data=request.data, partial=partial)
@@ -297,6 +382,16 @@ class PostsViewSet(viewsets.ModelViewSet):
         serializer = PostSerializer(instance=instance)
 
         return Response(serializer.data)
+    
+    def destroy(self, request, author_pk=None, *args, **kwargs):
+        user = request.user
+        author = Author.objects.get(id=author_pk)
+        if (author.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        instance = self.get_object()
+        self.perform_destroy(instance)
+        return Response(status=status.HTTP_204_NO_CONTENT)
     
     @action(detail=True)
     def image(self, request, author_pk, pk, *args, **kwargs):
@@ -315,25 +410,20 @@ class PostsViewSet(viewsets.ModelViewSet):
     @action(detail=True, methods=['Post'])
     def like(self, request, author_pk, pk, *args, **kwargs):
         object = self.get_object()
-        serializer = LikeSerializer(data=request.data)
-        if serializer.is_valid():
-            authorid = serializer.data.get('author')
-            author = get_object_or_404(Author, id=authorid)
-            summary = f"{author.displayName} liked your post"
-            if Likes.objects.filter(author=author, object_id=pk).count():
-                return Response({"detail": ["Request already exists."]},
-                            status=status.HTTP_400_BAD_REQUEST)
-            like = Likes(summary=summary, author=author, object=object)
-            like.save()
-            object.numLikes += 1
-            object.save()
-            liked = Liked.objects.get(author=author)
-            liked.items.add(like)
-            return Response({"detail": ["Liked post."]},
-                        status=status.HTTP_200_OK)
-        
-        return Response(serializer.errors,
+        user = request.user
+        author = Author.get_author_from_user(user=user)
+        summary = f"{author.displayName} liked your post"
+        if Likes.objects.filter(author=author, object_id=pk).count():
+            return Response({"detail": ["Request already exists."]},
                         status=status.HTTP_400_BAD_REQUEST)
+        like = Likes(summary=summary, author=author, object=object)
+        like.save()
+        object.numLikes += 1
+        object.save()
+        liked = Liked.objects.get(author=author)
+        liked.items.add(like)
+        return Response({"detail": ["Liked post."]},
+                    status=status.HTTP_200_OK)
     
     @action(detail=False)
     def public(self, request, author_pk=None, *args, **kwargs):
@@ -346,15 +436,30 @@ class PostsViewSet(viewsets.ModelViewSet):
     @action(detail=False)
     def feed(self, request, author_pk=None, *args, **kwargs):
         self.pagination_class=PostsPagination
-        following = Following.objects.get(author__id=author_pk)
-        followed_authors = following.items.all()
-        posts = Post.objects.filter(author__in=followed_authors)
+        author = Author.objects.get(id=author_pk)
+        followers = Followers.objects.filter(author=author).values_list('items', flat=True)
+        following = Following.objects.filter(author=author).values_list('items', flat=True)
+        friends = Author.objects.filter(
+            Q(id__in=followers) & Q(id__in=following)
+        ).values_list('id', flat=True)
+
+        my_posts = Post.objects.filter(Q(author__id=author_pk))
+        following_posts = Post.objects.filter(
+            Q(author__id__in=following) & ~Q(visibility='FRIENDS')
+        )
+        friend_posts = Post.objects.filter(
+            Q(author__id__in=friends) & Q(visibility='FRIENDS')
+        )
+        # public_posts = Post.objects.filter(visibility='PUBLIC')
+
+        posts = my_posts | following_posts | friend_posts
+
         page = self.paginate_queryset(posts)
         serializer = PostSerializer(page, many=True)
         return self.get_paginated_response(serializer.data)
 
 
-class CommentsViewSet(viewsets.ModelViewSet):
+class CommentsViewSet(viewsets.GenericViewSet):
     queryset = Comment.objects.all()
     serializer_class = CommentsSerializer
     pagination_class = CommentsPagination
@@ -390,8 +495,13 @@ class CommentsViewSet(viewsets.ModelViewSet):
         return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
     
     def update(self, request, *args, **kwargs):
+        user = request.user
         partial = kwargs.pop('partial', False)
         instance = self.get_object()
+        author = instance.author
+        if (author.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
         serializer = self.get_serializer(instance, data=request.data, partial=partial)
         serializer.is_valid(raise_exception=True)
         comment = serializer.validated_data['comment']
@@ -400,6 +510,16 @@ class CommentsViewSet(viewsets.ModelViewSet):
         serializer = CommentsSerializer(instance=instance)
 
         return Response(serializer.data)
+    
+    def destroy(self, request, *args, **kwargs):
+        user = request.user
+        instance = self.get_object()
+        author = instance.author
+        if (author.user != user):
+            return Response({"detail": ["Not authorized to do that."]},
+                            status=status.HTTP_401_UNAUTHORIZED)
+        self.perform_destroy(instance)
+        return Response(status=status.HTTP_204_NO_CONTENT)
     
     @action(detail=True)
     def likes(self, request, author_pk, post_pk, pk, *args, **kwargs):
